Selected piece did not move to destination square
There was a logic error that caused the selected piece to not move to a destination square if one of the rows and columns were the same (Ex: pawn from e2 to e4). The logic for that code has been corrected. Added functionality to all grids clicked on the chessboard (since the onClick() listener wasn't fully implemented, clicking on some items on the grid with a missing onClick() listener method would cause the entire app to crash, this has been fixed by adding all onClick() listeners to all ImageView grid items in the PlayComputer.java file